#60 Enhance TextDisplay and WordHighlighter components with containerRef for improved scrolling and layout management
