+++ src/features/lireFlow/components/WordHighlighter.tsx
@@ -1,8 +1,5 @@
-import { useEffect, useState, useRef, RefObject } from 'react';
-import { hyphenate } from 'hyphen/pt';
-import { averageSyllableTime } from './readingParameters';
-import { PUNCTUATION_MARKS_TIME } from '../constants/readingContants';
-import { RegexConstants } from '../constants/regexConstants';
+import { RefObject, useEffect } from 'react';
+import { useWordHighlighter } from '../hooks/useWordHighlighter';
 
 const WordHighlighter = ({
   paragraph,
@@ -19,157 +16,25 @@ const WordHighlighter = ({
   wordsPerMinuteRef: RefObject<number>;
   containerRef: RefObject<HTMLDivElement | null>;
 }) => {
-  const [currentIndex, setCurrentIndex] = useState<number | null>(null);
-  const elementIndexs = useRef<number[]>([]);
-  const indexRef = useRef<number>(0);
-  const isReadingRef = useRef<boolean>(isReading);
-  const timeoutRef = useRef<number | null>(null);
-  const currentWordRef = useRef<HTMLSpanElement | null>(null);
+  const {
+    elements,
+    currentIndex,
+    currentWordRef,
+    initializeElementIndexs,
+    runReadingFlow,
+    scrollToCurrentWord,
+  } = useWordHighlighter({
+    paragraph,
+    onFinish,
+    isReading,
+    speedRef,
+    wordsPerMinuteRef,
+    containerRef,
+  });
 
-  const elements: string[] = paragraph.split(/(\s+|[^\wÀ-ÖØ-öø-ÿ])/);
-
-  const isWord = (element: string): boolean => /^[\wÀ-ÖØ-öø-ÿ]+$/.test(element);
-  const isPunctuation = (element: string): boolean =>
-    /^[.,!?;:"()]+$/.test(element);
-
-  useEffect((): void => {
-    elementIndexs.current = elements
-      .map((element: string, index: number): number | null =>
-        isWord(element) || isPunctuation(element) ? index : null,
-      )
-      .filter((index): index is number => index !== null);
-
-    indexRef.current = 0;
-    setCurrentIndex(null);
-  }, [paragraph]);
-
-  useEffect((): (() => void) => {
-    isReadingRef.current = isReading;
-
-    const fixWordHyphenation = (word: string, hyphenated: string): string => {
-      if (word.length <= 2) return word;
-
-      const {
-        vowel,
-        startWithVowel,
-        endWithTwoVowels,
-        strongVowels,
-        weakVowels,
-      } = RegexConstants;
-
-      if (startWithVowel.test(word)) {
-        const fixHyphenationWordsStartWithVowel = (hyphenated: string) => {
-          const shouldAttachConsonantToPreviousVowel =
-            /^[lmnrs]$/i.test(word[1]) && !vowel.test(word[2]);
-
-          if (shouldAttachConsonantToPreviousVowel) {
-            return hyphenated.slice(0, 2) + '-' + hyphenated.slice(2);
-          }
-
-          return hyphenated.slice(0, 1) + '-' + hyphenated.slice(1);
-        };
-
-        hyphenated = fixHyphenationWordsStartWithVowel(hyphenated);
-      }
-
-      if (endWithTwoVowels.test(word)) {
-        const applyHiatusHyphenationRule = (hyphenated: string) => {
-          const lastIndex = word.length - 1;
-          const lastLetter = word[lastIndex];
-          const penulLetter = word[lastIndex - 1];
-          const isHiatus =
-            (strongVowels.test(penulLetter) && strongVowels.test(lastLetter)) ||
-            (strongVowels.test(penulLetter) && weakVowels.test(lastLetter));
-
-          if (isHiatus) {
-            return (
-              hyphenated.slice(0, lastIndex) + '-' + hyphenated.slice(lastIndex)
-            );
-          }
-
-          return word;
-        };
-        hyphenated = applyHiatusHyphenationRule(hyphenated);
-      }
-
-      if (!hyphenated.includes('-') && word.length > 3) {
-        hyphenated = word.slice(0, 2) + '-' + word.slice(2);
-      }
-
-      return hyphenated;
-    };
-
-    const calculateWordTime = async (word: string): Promise<number> => {
-      const hyphenatedText: string = fixWordHyphenation(
-        word,
-        await hyphenate(word, { hyphenChar: '-' }),
-      );
-
-      const syllablesCount: number = hyphenatedText
-        .split('-')
-        .filter((syllable) => syllable.trim() !== '').length;
-      return Math.round(
-        (syllablesCount * averageSyllableTime(wordsPerMinuteRef.current)) /
-          speedRef.current,
-      );
-    };
-
-    const highlightFlow = async (): Promise<void> => {
-      while (indexRef.current < elementIndexs.current.length) {
-        if (!isReadingRef.current) return;
-
-        const element: string =
-          elements[elementIndexs.current[indexRef.current]];
-
-        let waitTime: number = 0;
-
-        if (isWord(element)) {
-          setCurrentIndex(elementIndexs.current[indexRef.current]);
-          waitTime = await calculateWordTime(element);
-        } else {
-          waitTime =
-            PUNCTUATION_MARKS_TIME.find((mark) => mark.mark === element)
-              ?.time || 150;
-        }
-
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, waitTime)),
-        );
-
-        indexRef.current++;
-      }
-
-      setCurrentIndex(null);
-      if (onFinish) {
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, 500)),
-        );
-        onFinish();
-      }
-    };
-
-    highlightFlow();
-
-    return () => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current);
-      }
-    };
-  }, [isReading]);
-
-  useEffect(() => {
-    if (!currentWordRef.current || !containerRef.current) return;
-
-    const word = currentWordRef.current;
-    const container = containerRef.current;
-
-    const target =
-      word.offsetTop - container.clientHeight / 4 + word.offsetHeight;
-
-    if (Math.abs(container.scrollTop - target) > 4) {
-      container.scrollTo({ top: target, behavior: 'smooth' });
-    }
-  }, [currentIndex]);
+  useEffect(initializeElementIndexs, [initializeElementIndexs]);
+  useEffect(runReadingFlow, [runReadingFlow]);
+  useEffect(scrollToCurrentWord, [scrollToCurrentWord]);
 
   return elements.map((element: string, index: number) => (
     <span
