+++ src/features/lireFlow/hooks/useWordHighlighter.ts
@@ -1,9 +1,8 @@
-import { useState, useRef, RefObject, useCallback, useMemo } from 'react';
+import { useCallback, useState, useRef, RefObject } from 'react';
 import { hyphenate } from 'hyphen/pt';
-import {
-  averageSyllableTime,
-  punctuationMarksTime,
-} from '../components/readingParameters';
+import { averageSyllableTime } from '../components/readingParameters';
+import { PUNCTUATION_MARKS_TIME } from '../constants/readingContants';
+import { RegexConstants } from '../constants/regexConstants';
 
 export const useWordHighlighter = ({
   paragraph,
@@ -11,12 +10,14 @@ export const useWordHighlighter = ({
   isReading,
   speedRef,
   wordsPerMinuteRef,
+  containerRef,
 }: {
   paragraph: string;
   onFinish?: () => void;
   isReading: boolean;
   speedRef: RefObject<number>;
   wordsPerMinuteRef: RefObject<number>;
+  containerRef: RefObject<HTMLDivElement | null>;
 }) => {
   const [currentIndex, setCurrentIndex] = useState<number | null>(null);
   const elementIndexs = useRef<number[]>([]);
@@ -25,16 +26,14 @@ export const useWordHighlighter = ({
   const timeoutRef = useRef<number | null>(null);
   const currentWordRef = useRef<HTMLSpanElement | null>(null);
 
-  const wordsAndSeparators = useMemo(() => {
-    return paragraph.split(/(\s+|[^\wÀ-ÖØ-öø-ÿ])/);
-  }, [paragraph]);
+  const elements: string[] = paragraph.split(/(\s+|[^\wÀ-ÖØ-öø-ÿ])/);
 
   const isWord = (element: string): boolean => /^[\wÀ-ÖØ-öø-ÿ]+$/.test(element);
   const isPunctuation = (element: string): boolean =>
     /^[.,!?;:"()]+$/.test(element);
 
-  const initializeIndexes = useCallback(() => {
-    elementIndexs.current = wordsAndSeparators
+  const initializeElementIndexs = useCallback((): void => {
+    elementIndexs.current = elements
       .map((element: string, index: number): number | null =>
         isWord(element) || isPunctuation(element) ? index : null,
       )
@@ -42,54 +41,112 @@ export const useWordHighlighter = ({
 
     indexRef.current = 0;
     setCurrentIndex(null);
-  }, [wordsAndSeparators]);
+  }, [paragraph]);
 
-  const updateReadingState = useCallback(() => {
-    isReadingRef.current = isReading;
+  const fixWordHyphenation = (word: string, hyphenated: string): string => {
+    if (word.length <= 2) return word;
 
-    const calculateWordTime = async (word: string): Promise<number> => {
-      const hyphenatedText: string = await hyphenate(word, { hyphenChar: '-' });
-      const syllablesCount: number = hyphenatedText.split('-').length;
-      const variableNameWithAnMeaning =
-        (syllablesCount * averageSyllableTime(wordsPerMinuteRef.current)) /
-        speedRef.current;
-      return Math.round(variableNameWithAnMeaning);
-    };
+    const {
+      vowel,
+      startWithVowel,
+      endWithTwoVowels,
+      strongVowels,
+      weakVowels,
+    } = RegexConstants;
 
-    const highlightFlow = async (): Promise<void> => {
-      while (indexRef.current < elementIndexs.current.length) {
-        if (!isReadingRef.current) return;
+    if (startWithVowel.test(word)) {
+      const fixHyphenationWordsStartWithVowel = (hyphenated: string) => {
+        const shouldAttachConsonantToPreviousVowel =
+          /^[lmnrs]$/i.test(word[1]) && !vowel.test(word[2]);
 
-        const element: string =
-          wordsAndSeparators[elementIndexs.current[indexRef.current]];
-
-        let waitTime: number = 0;
-
-        if (isWord(element)) {
-          setCurrentIndex(elementIndexs.current[indexRef.current]);
-          waitTime = await calculateWordTime(element);
-        } else {
-          waitTime =
-            punctuationMarksTime.find(
-              (mark: { mark: string }) => mark.mark === element,
-            )?.time || 150;
+        if (shouldAttachConsonantToPreviousVowel) {
+          return hyphenated.slice(0, 2) + '-' + hyphenated.slice(2);
         }
 
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, waitTime)),
-        );
+        return hyphenated.slice(0, 1) + '-' + hyphenated.slice(1);
+      };
 
-        indexRef.current++;
+      hyphenated = fixHyphenationWordsStartWithVowel(hyphenated);
+    }
+
+    if (endWithTwoVowels.test(word)) {
+      const applyHiatusHyphenationRule = (hyphenated: string) => {
+        const lastIndex = word.length - 1;
+        const lastLetter = word[lastIndex];
+        const penulLetter = word[lastIndex - 1];
+        const isHiatus =
+          (strongVowels.test(penulLetter) && strongVowels.test(lastLetter)) ||
+          (strongVowels.test(penulLetter) && weakVowels.test(lastLetter));
+
+        if (isHiatus) {
+          return (
+            hyphenated.slice(0, lastIndex) + '-' + hyphenated.slice(lastIndex)
+          );
+        }
+
+        return word;
+      };
+      hyphenated = applyHiatusHyphenationRule(hyphenated);
+    }
+
+    if (!hyphenated.includes('-') && word.length > 3) {
+      hyphenated = word.slice(0, 2) + '-' + word.slice(2);
+    }
+
+    return hyphenated;
+  };
+
+  const calculateWordTime = async (word: string): Promise<number> => {
+    const hyphenatedText: string = fixWordHyphenation(
+      word,
+      await hyphenate(word, { hyphenChar: '-' }),
+    );
+
+    const syllablesCount: number = hyphenatedText
+      .split('-')
+      .filter((syllable) => syllable.trim() !== '').length;
+
+    return Math.round(
+      (syllablesCount * averageSyllableTime(wordsPerMinuteRef.current)) /
+        speedRef.current,
+    );
+  };
+
+  const highlightFlow = async (): Promise<void> => {
+    while (indexRef.current < elementIndexs.current.length) {
+      if (!isReadingRef.current) return;
+
+      const element: string = elements[elementIndexs.current[indexRef.current]];
+
+      let waitTime: number = 0;
+
+      if (isWord(element)) {
+        setCurrentIndex(elementIndexs.current[indexRef.current]);
+        waitTime = await calculateWordTime(element);
+      } else {
+        waitTime =
+          PUNCTUATION_MARKS_TIME.find((mark) => mark.mark === element)?.time ||
+          150;
       }
 
-      setCurrentIndex(null);
-      if (onFinish) {
-        await new Promise(
-          (resolve) => (timeoutRef.current = setTimeout(resolve, 500)),
-        );
-        onFinish();
-      }
-    };
+      await new Promise(
+        (resolve) => (timeoutRef.current = setTimeout(resolve, waitTime)),
+      );
+
+      indexRef.current++;
+    }
+
+    setCurrentIndex(null);
+    if (onFinish) {
+      await new Promise(
+        (resolve) => (timeoutRef.current = setTimeout(resolve, 500)),
+      );
+      onFinish();
+    }
+  };
+
+  const runReadingFlow = useCallback((): (() => void) => {
+    isReadingRef.current = isReading;
 
     highlightFlow();
 
@@ -98,22 +155,34 @@ export const useWordHighlighter = ({
         clearTimeout(timeoutRef.current);
       }
     };
-  }, [isReading, onFinish, wordsAndSeparators, speedRef, wordsPerMinuteRef]);
+  }, [
+    isReading,
+    elements,
+    onFinish,
+    wordsPerMinuteRef.current,
+    speedRef.current,
+  ]);
 
   const scrollToCurrentWord = useCallback(() => {
-    if (currentWordRef.current) {
-      currentWordRef.current.scrollIntoView({
-        behavior: 'smooth',
-        block: 'nearest',
-      });
+    if (!currentWordRef.current || !containerRef.current) return;
+
+    const word = currentWordRef.current;
+    const container = containerRef.current;
+
+    const target =
+      word.offsetTop - container.clientHeight / 4 + word.offsetHeight;
+
+    if (Math.abs(container.scrollTop - target) > 4) {
+      container.scrollTo({ top: target, behavior: 'smooth' });
     }
-  }, []);
+  }, [currentIndex, currentWordRef, containerRef]);
+
   return {
-    wordsAndSeparators,
+    elements,
     currentIndex,
     currentWordRef,
-    initializeIndexes,
-    updateReadingState,
+    initializeElementIndexs,
+    runReadingFlow,
     scrollToCurrentWord,
   };
 };
